hive-backend-alpha/app/core/auth.py
from typing import Optional
from uuid import UUID
from jose import JWTError, jwt
from app.config import settings


def verify_access_token(token: str) -> Optional[UUID]:
    try:
        payload = jwt.decode(token, settings.jwt_secret_key, algorithms=[settings.jwt_algorithm])
        user_id: str = payload.get("sub")
        if user_id is None:
            return None
        return UUID(user_id)
    except (JWTError, ValueError):
        return None